:heavy_minus_sign: Remove puppeteer dependency
Fix problem with CSR
