+++ package.json
@@ -30,7 +30,6 @@
     "hono": "2.7.1",
     "husky": "8.0.3",
     "picocolors": "1.0.0",
-    "puppeteer": "19.4.1",
     "tailwindcss": "3.2.4"
   },
   "devDependencies": {
